backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Prelude Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.135.0",
    "uvicorn[standard]>=0.37.0",
    "sqlalchemy>=2.0.44",
    "pydantic>=2.12.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.22",
    "pymysql>=1.1.2",
    "bcrypt>=4.2.0",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.4",
]

[tool.fastapi]
entrypoint = "app.main:app"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff.lint.per-file-ignores]
# Ignore E402 (module level import not at top of file) in test fixtures
# This is necessary because conftest.py must load environment variables before imports
"tests/conftest.py" = ["E402"]
